Don't include the reason for rejectionhandled
In Chrome, we'd have to keep a weak reference to the reason to avoid leaking memory, but then it's sometimes null and sometimes not. This is due to bugs in our Blink vs. V8 memory management. See https://crbug.com/501866.

Instead we just leave it always null. Authors can keep a side-table of promise => reason if they want it; usually it won't be necessary, since rejectionhandled is mostly used for correlating back to previous unhandledrejection events.

Discussion in #7. If other vendors object this can be revisited.